Drop unnecessary logic from PopScanQ.populate
Checking the types of function parameters is usually unnecessary
and potentially confusing readability wise, since the function
is could/would be used in such many different ways.

Of course there are exceptions for this. Some really flexible
APIs makes good use of this kind of behaviour.

In this case it seems a lot easier for caller to provide a list
and there is no need to make further checks or conversions. KISS:)